Move most of variable list into elements file
